Fall back to shader when subsurface cannot blit
glBlitFramebuffer is a GLES3 / OpenGL 3.0 function that may be missing or
broken on some drivers. When the subsurface's EGL context cannot use it,
draw the engine frame to the window surface with the compositor shader
instead, matching the fallback the OpenGL renderer already uses.

The glBlitFramebuffer availability check (driver vendor blocklist plus GL
version / GL_EXT_framebuffer_blit) is moved from FlCompositorOpenGL into
fl_opengl_manager_can_blit so both renderers share the same detection.
